Move nerdtree into a file browser plugin section and reformat
