+++ lua/warpcode/plugins/plugin/file-browser.lua
@@ -1,0 +1,78 @@
+local pexec = require('warpcode.utils.priority-exec')
+
+return {
+  'scrooloose/nerdtree',
+  dependencies = {
+    { 'Xuyuanp/nerdtree-git-plugin' },
+    { 'jistr/vim-nerdtree-tabs' },
+  },
+  config = function()
+    vim.cmd([[
+        let g:NERDTreeChDirMode=2
+        let g:NERDTreeIgnore=['\.rbc$', '\~$', '\.pyc$', '\.db$', '\.sqlite$', '__pycache__']
+        let g:NERDTreeSortOrder=['^__\.py$', '\/$', '*', '\.swp$', '\.bak$', '\~$']
+        let g:NERDTreeShowBookmarks=1
+        let g:nerdtree_tabs_focus_on_files=1
+        let g:NERDTreeMapOpenInTabSilent = '<RightMouse>'
+        let g:NERDTreeWinSize = 40
+        let g:NERDTreeShowHidden=1
+        " Show ignored icon next to ignored files. This can be intensive with a lot of files
+        let g:NERDTreeGitStatusShowIgnored = 1
+        " List of icons for file statuses
+        let g:NERDTreeGitStatusIndicatorMapCustom = {
+                    \ "Modified"  : "✹",
+                    \ "Staged"    : "✚",
+                    \ "Untracked" : "✭",
+                    \ "Renamed"   : "➜",
+                    \ "Unmerged"  : "═",
+                    \ "Deleted"   : "✖",
+                    \ "Dirty"     : "✗",
+                    \ "Clean"     : "✔︎",
+                    \ 'Ignored'   : '☒',
+                    \ "Unknown"   : "?"
+                    \ }
+
+        " NERDTree - Quit vim when all other windows have been closed.
+        au BufEnter *
+                    \ if (winnr("$") == 1 && exists("b:NERDTreeType") &&
+                    \ b:NERDTreeType == "primary") |
+                    \   q |
+                    \ endif
+
+        "
+        " Auto find files in nerdtree on buf enter
+        "
+        " Check if NERDTree is open or active
+        function! IsNERDTreeOpen()
+            return exists("t:NERDTreeBufName") && (bufwinnr(t:NERDTreeBufName) != -1)
+        endfunction
+
+        " Call NERDTreeFind if NERDTree is active, current window contains a modifiable
+        " file, and we're not in vimdiff
+        function! SyncTree()
+            if &modifiable && IsNERDTreeOpen() && strlen(expand('%')) > 0 && !&diff
+                NERDTreeFind
+                wincmd p
+            endif
+        endfunction
+
+        " Highlight currently open buffer in NERDTree
+        " autocmd BufEnter * call SyncTree()
+
+        " With the auto find feature, ensure we don't double open nerdtree
+        function! ToggleNerdTree()
+            set eventignore=BufEnter
+            NERDTreeToggle
+            set eventignore=
+        endfunction
+        nnoremap <leader>t :call ToggleNerdTree()<CR>
+        ]])
+
+    pexec.addCall('fs.file_tree', function()
+      vim.cmd('NERDTreeToggle')
+    end, 10)
+    pexec.addCall('fs.find_buffer', function()
+      vim.cmd('NERDTreeFind')
+    end, 10)
+  end,
+}
